Fixing sidebar render for logged in user
